Rework the page toolbar to override it rather than hacking it via a monkey patch, this gives versioning much more control over the toolbar
