+++ djangocms_versioning/monkeypatch/__init__.py
@@ -1,7 +1,6 @@
 from . import (  # noqa: F401
     admin,
     checks,
-    cms_toolbars,
     extensions,
     menu,
     page,
